Modify endpointing for Deepgram
Adding a larger default value for endpointing for Deepgram to use when
determining when to set `speech_final`. The original default is 10 ms.
Also require some silence, set by `time_cutoff_seconds`, in addition to
a punctuation character when using `PunctuationEndpointingConfig`.

These changes prevent the agent from trying to talk over the caller as
much, and then getting confused. It may result in a longer pause before
the agent responds, though.

See https://developers.deepgram.com/docs/endpointing and
https://developers.deepgram.com/docs/understand-endpointing-interim-results
for details on the Deepgram side.